misc command updates

+++ bot/commands/ping.js
@@ -1,6 +1,9 @@
-/* todo: handle deletion when the bot cant delete. since this is such a simple command, i just barred it via clientPermissions, but i want the user to still be able to do it in a channel where the bot'd otherwise be unable to delete their message, you feel me bro */
-
 const CommandBlock = require("../../modules/CommandBlock");
+const { WebhookClient } = require("discord.js");
+
+/** @todo May want to look into how the bot reacts when it cant `message.delete()` or `message.react()` */
+
+const ping = "<a:_:597509670210633729>";
 
 module.exports = [
     new CommandBlock({
@@ -9,19 +12,45 @@ module.exports = [
         description: "Two latency statistics, the rough time it took to respond and the bot's average heartbeat. Generally used to check if the bot is responsive.",
         clientPermissions: ["VIEW_CHANNEL", "SEND_MESSAGES"],
     }, async function(client, message, content, args) {
-        const msg = await message.channel.send("<a:_:597509670210633729>");
+        const msg = await message.channel.send(ping);
         msg.edit(`\uD83C\uDFD3 Pong!\nResponse time is \`${msg.createdTimestamp - message.createdTimestamp}ms\`.`);
     }),
     new CommandBlock({
         identity: ["echo", "e"],
         description: "Echoes text.",
-        usage: "[text]",
-        clientPermissions: ["VIEW_CHANNEL", "SEND_MESSAGES", "MANAGE_MESSAGES"],
+        usage: "[...text]",
+        clientPermissions: ["VIEW_CHANNEL", "SEND_MESSAGES"],
     }, async function(client, message, content, args) {
-        await message.delete();
-        return message.channel.send({
+        await message.channel.send({
             content: content,
             allowedMentions: { parse: [] },
         });
+        return message.delete();
     }),
+    new CommandBlock({
+        identity: ["wecho"],
+        description: "Echoes text via a Discord webhook.\n\n**Note:** This is not secure; if used publicly, people WILL be able to get your webhooks ID and token, which will let them use it aswell! Use only as a utility!",
+        usage: "[hook_url] [...text]",
+        clientPermissions: ["VIEW_CHANNEL", "SEND_MESSAGES", "MANAGE_MESSAGES"],
+        locked: ["hosts"],
+    }, async function(client, message, content, [hook_url, ...args]) {
+        const negative = client.config.get("metadata.reactions.negative").value();
+        
+        const matched = hook_url.match(/discord.com\/api\/webhooks\/([^\/]+)\/([^\/]+)/);
+        if(!matched[1] || !matched[2]) {
+            return message.channel.send(`<:_:${negative}> Invalid webhook url provided.`);
+        }
+
+        try {
+            const hook = new WebhookClient(matched[1], matched[2]);
+            await hook.send({
+                content: args.join(" "),
+                allowedMentions: { parse: [] },
+            });
+            return message.delete();
+        } catch(e) {
+            message.react(negative);
+            return message.channel.send(`<:_:${negative}> An error occured;\`\`\`\n${e}\`\`\``);
+        }
+    })
 ];
